rename partial penalty payment storage to `partialPenaltyPayment`
and add clarifying comments around the partial payment logic.

the previous naming was confusing. In the scenario where repayment
covers only part of the penalty debt, we need to track the partial
payment in order to compute the remaining penalty debt. When
penalty debt is fully paid, we set the partial payment amount
to 0