use compound interest rate
before this change, linear interest rate formula was used, but
interest would be compounded whenever _updateAccumulatedInterestFactor()
was called. This resulted in a potentially different debt amount
for borrowers with the same interest rate over the same time
duration.

Changing to the compound rate formula ensures that interest is
continuously compounded, so it will be consistent regardless of the
frequency of accumulatedInterestFactor updates.